+++ src/tests/e2e/auth.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@/tests/e2e/fixtures/auth.fixture'
-import resetDb from './utils/resetdb'
+import resetDb from './utils/resete2edb'
 
 test.describe("test auth flow", () => {
     test.afterEach(async () => {
