+++ src/tests/e2e/home.spec.ts
@@ -1,0 +1,2 @@
+import { test, expect } from '@/tests/e2e/fixtures/auth.fixture'
+import resetDb from './utils/resete2edb'
